requirements.txt
streamlit
crewai
crewai_tools
python-dotenv
-e.